bring in upstream change to Fripon.py

+++ wmpl/Formats/Fripon.py
@@ -40,7 +40,7 @@ FRIPON_STATION_FILE = "FRIPON_location_list.csv"
 
 
 
-def loadFripon(dir_path, fripon_paths):
+def loadFripon(dir_path, fripon_paths, overwrite_fripon_stations=False):
 
 
     ### Download FRIPON station data from the web, or load it locally if available ###
@@ -48,22 +48,25 @@ def loadFripon(dir_path, fripon_paths):
     # Local path to the file where it will be saved
     station_file_path = os.path.join(dir_path, FRIPON_STATION_FILE)
 
-    try:
-        # Download the FRIPON station files
-        print("Downloading FRIPON station file from:", FRIPON_STATIONS_URL)
-        urllibrary.urlretrieve(FRIPON_STATIONS_URL, station_file_path)
-    except:
+    # If a file available locally, use it unless the overwrite flag is set
+    if not os.path.exists(station_file_path) or overwrite_fripon_stations:
 
-        # If the file cannot be downloaded, try loading it locally
-        print("The FRIPON station file could not be retrieved!")
-        
+        try:
+            # Download the FRIPON station files
+            print("Downloading FRIPON station file from:", FRIPON_STATIONS_URL)
+            urllibrary.urlretrieve(FRIPON_STATIONS_URL, station_file_path)
+        except:
 
-        if os.path.exists(station_file_path):
-            print("Loading a local file...")
+            # If the file cannot be downloaded, try loading it locally
+            print("The FRIPON station file could not be retrieved!")
+            
 
-        else:
-            print("A local station file could not be found:", station_file_path)
-            sys.exit("Could not read station data")
+            if os.path.exists(station_file_path):
+                print("Loading a local file...")
+
+            else:
+                print("A local station file could not be found:", station_file_path)
+                sys.exit("Could not read station data")
 
 
     ###
@@ -190,6 +193,10 @@ if __name__ == "__main__":
     arg_parser.add_argument('-w', '--walk', \
         help="Recursively find FRIPON .met files in the given folder and use them for trajectory estimation.", \
         action="store_true")
+    
+    arg_parser.add_argument('--updatestations', \
+        help="Force update of the FRIPON station data file.", \
+        action="store_true")
 
     # Parse the command line arguments
     cml_args = arg_parser.parse_args()
@@ -264,7 +271,8 @@ if __name__ == "__main__":
 
 
     # Load the observations into container objects
-    jdt_ref, meteor_list = loadFripon(dir_path, fripon_paths)
+    jdt_ref, meteor_list = loadFripon(dir_path, fripon_paths, 
+                                      overwrite_fripon_stations=cml_args.updatestations)
 
 
     # Save the data as RMS JSON files
